+++ src/FluxFlow.Components.Payloads.Composition/PayloadsCompositionNodeRegistryExtensions.cs
@@ -2,6 +2,7 @@ using FluxFlow.Components.Payloads.Contracts;
 using FluxFlow.Components.Payloads.Nodes;
 using FluxFlow.Components.Payloads.Options;
 using FluxFlow.Composition;
+using FluxFlow.Data;
 
 namespace FluxFlow.Components.Payloads.Composition;
 
@@ -19,12 +20,12 @@ public static class PayloadsCompositionNodeRegistryExtensions
             CreatePayloadInspectNode,
             inputs:
             [
-                CompositionPorts.Metadata<PayloadInspectionRequest>(
+                CompositionPorts.Metadata<FlowContent>(
                     PayloadsCompositionPortNames.Input)
             ],
             outputs:
             [
-                CompositionPorts.Metadata<PayloadInspectionResult>(
+                CompositionPorts.Metadata<FlowResult<PayloadInspectionResult>>(
                     PayloadsCompositionPortNames.Output)
             ]);
     }
@@ -33,25 +34,26 @@ public static class PayloadsCompositionNodeRegistryExtensions
         CompositionNodeFactoryContext context)
     {
         var options = context.BindConfiguration<PayloadInspectOptions>();
+        var codecs = context.GetResource<FlowContentCodecCatalog>(
+            PayloadsCompositionResourceNames.Codecs);
         var clock = context.GetResource<TimeProvider>(
             PayloadsCompositionResourceNames.Clock);
-        var node = new PayloadInspectNode(options, clock);
+        var node = new FlowContentInspectNode(options, codecs, clock);
 
         return ValueTask.FromResult(ComposedNode.Create(
             node,
             inputs:
             [
-                CompositionPorts.Input<PayloadInspectionRequest>(
+                CompositionPorts.Input<FlowContent>(
                     PayloadsCompositionPortNames.Input,
                     node.Input)
             ],
             outputs:
             [
-                CompositionPorts.Output<PayloadInspectionResult>(
+                CompositionPorts.Output<FlowResult<PayloadInspectionResult>>(
                     PayloadsCompositionPortNames.Output,
                     node.Output)
             ],
-            events: node.Events,
-            errors: node.Errors));
+            events: node.Events));
     }
 }