+++ src/FluxFlow.Components.Payloads.Composition/PayloadsCompositionResourceNames.cs
@@ -2,5 +2,7 @@ namespace FluxFlow.Components.Payloads.Composition;
 
 public static class PayloadsCompositionResourceNames
 {
+    public const string Codecs = "codecs";
+
     public const string Clock = "clock";
 }